Version 1 1

+++ database/tables/user_login/auth.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS auth(
 id SERIAL PRIMARY KEY,
 hash TEXT NOT NULL,
-auth_time int REFERENCES "time"(id) NOT NULL
+auth_time bigint REFERENCES "time"(id) NOT NULL
 );
